+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ### Added
 - License, initial mod setup, mod metadata etc
 - Block highlighting util
+- simple config manager